backend/requirements.txt
astroid==1.6.2
click==6.7
cycler==0.10.0
decorator==4.2.1
Flask==0.12.2
Flask-Script==2.0.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
networkx==2.1
numpy==1.14.2
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.7.1
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
